+++ syncalgorithms/graceperiodnetworkdelaysyncalgo.cpp
@@ -77,7 +77,7 @@ namespace sim_comm{
       {
 	  if(busywait)
 	    this->interface->waitforAll();
-          uint8_t diff=interface->reduceTotalSendReceive();
+          uint64_t diff=interface->reduceTotalSendReceive();
           //network unstable, we need to wait!
           nextEstTime=currentTime+convertToFrameworkTime(Integrator::getCurSimMetric(),1); 
 	  minnetworkdelay=interface->reduceNetworkDelay();
